Fixed bug causing server to make connections per requests instead of using just one connection
